// src/app/Model/subclient.ts
export class Subclient {
  id: number;
  subclientid: string;
  subclientcode: string;
  subclientdivisioncode: string;
  subclientname: string;
  subclientcontactname: string;
  subclientcontactemail: string;
  subclientcontactphone: string;
  subclientstatus: boolean;
  displayname: string;
  subclientsplitparameters: any;
  subclientoutboundftp: any;
}