// src/app/Model/Configuration/providermapping.ts
export class Providermapping {
  mappingid: string;
  providernpi: string;
  subclientId: string;
  groupid: string;
  subgroupid: string;
  firstname: string;
  middlename: string;
  lastname: string;
  sname: string;
  mailingaddress: string;
  npitype: string;
  gender: string;
  dob: string;
  status: string;
  createdon: string;
  createdby: string;
  subclientcode: string;
  subclientname: string;
  clientid: string;
}