// src/app/Model/client.ts
import { Ftpmaster } from './ftpmaster';

export class Client {
    clientid: string;
    clientcode: string;
    clientname: string;
    clientcontactname: string;
    clientcontactemail: string;
    clientcontactphone: string;
    clientstatus: string;
    clientservicetypes: any;
    clientinboundftp: any;
    subclients: any;
}